Fix posix_find_parent to differentiate when /fs1 and /fs10 are filesystems
If we are finding the parent of /fs10/fs2, we don't want /fs1 to be the
parent. When we match "/fs1" against "/fs10/fs2", the strncmp will match,
so we test if the next character of the child patch is '/'. Since it's not,
we skip /fs1 as being a possible parent of /fs10/fs2.

Change-Id: I866d12f0f4ea32e2a498ed1fdd71e7974e2b249a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
(cherry picked from commit 25a549938c3ad361befa4177d9e06b2b949204c6)