Replace checkpatch.pl by a newer version (from kernel source).
This commit updates checkpatch.pl, replacing it with a newer version.
Please not that checkpatch.conf needs to be updated to, for checkpatch.pl
is now aware of Gerrit's changeids ("--ignore GERRIT_CHANGE_ID" is required).

Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>

Also, pull in spelling.txt from kernel.

Not being in the kernel, we don't have ARRAY_SIZE macro so some places
define it so we ignore ARRAY_SIZE.

Our coding style has too many if-then-else that have return in the middle
and thus create unnecessary else so ignore UNNECESSARY_ELSE.

Change-Id: Id57a022ea743e93dccf1d2bd76789bdc1c5c94aa
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>