FSAL_HPSS: brings FSAL back from its ashes.
FSAL_HPSS is now a glue to the FSAL CLAPI with no more adherence
to HPSS related code.

Please be aware of the following disclaimer:

Linkage/compilation/binding/loading/etc of HPSS licensed Software
must occur at the HPSS partner's or licensee's location.
It is not allowed to distribute this software as compiled or linked
binaries or libraries, as they include HPSS licensed material

Change-Id: Iacb2037fe30efe3094e4bf608979bc0c1602b049
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>
Signed-off-by: Dominique Martinet <dominique.martinet@cea.fr>