Lock TEST and UNLOCK operations will need a state_t to pass to FSAL
For some FSALs with multiple-fd support, the state_t will be associated
with what will be used to distinguish lock owners, and thus, if there
is a state_t mapping this cache entry and owner, then we must use it
so the FSAL will be able to make this association.

For example, FSAL_VFS attaches an file descriptor to each lock owner,
this is used with Open File Description Locks, so the fcntl MUST occur
on that file descriptor.

Since state_unlock() does not look at actual lock entries, we need to
pass in a state_t for this purpose.

In the case of an NLM_UNLOCK that comes in with no locks present, no
state_t will be found, that actually indicates that no locks are
present and we can ignore the unlock request.

Change-Id: Ia14825c49533fcc094464fb8196722df523db469
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>