Add zombie process detection on startup.
It is possible that ganesha threads hang in exit()
while waiting for file handles to be closed.
This results in <defunct> threads preventing a clean
restart of ganesha. This patch detects such a situation,
preventing a restart and logs the call stack of the
threads in question.

Change-Id: I51565be915d51b4a2fe8885d9f766df986c39f2a
Signed-off-by: Swen Schillig <swen@vnet.ibm.com>
(cherry picked from commit 09466ed83ecde31628e6b696ac6c517e1d511e4e)