Fix nfs4_compare_owner to not equate lock and open owners
There was some old code that when a lock owner was compared to an
open owner, would compare the open owner associated with the lock
owner to the open owner. The reason that was there is lost to time,
it may have been there due to some other bug to make things sort of
work (pass all known test cases...).

In any case, the code is not necessary and pynfs and cthon04 tests
pass just fine without it.

This mechanism was breaking the hash table of stateid by owner and entry.

Change-Id: Ifc041a1c082fd8628d21a8e559c75c20332c0f4d
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>