Log and abort inside memory allocation if it fails
Each function foo is renamed foo__ so that a macro foo can supply
the source file, line, and function.

The pool infrastructure passes these down.

After discussion, we have agreed to allow malloc(0) and calloc(0, 0) to
return non-NULL pointer, but we will expect that. Add checks in main
to test this assumption and exit Ganesha with an informative message if
the assumption is wrong.

Change-Id: I8e732dd8ea994bdaf0c63f395496915101a7c0e7
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>