Simplify abstract_mem pools
We currently do not use all the flexibility built in and
pool_basic_substrate is causing compiler warnings on gcc6.

Rename pool_init to pool_basic_init in deference to
possible future expansion.

In the meantime, we don't use constructors, so simplify
them out, and get rid of the substrate.

In the future, if we expand these, add the flexibility that
makes sense at that time.

Change-Id: I30977a78a5bf2f1be23112743ec0193cf1b948f8
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>