[nfs4_clientid.c] remove stale client_id reference from client_record
Experienced a race condition between the reaper-thread and another thread,
trying to reuse a formerly known/confirmed clientid record.
On IP move, nfs_release_v4_client() is setting the clientid to STALE
to indicate that it cannot be used anymore.
Later the call to nfs_client_id_get() changes the STALE state to EXPIRED
but does NOT remove the reference to the clientid record from
the client_record.
This is done by the reaper thread at some stage later, however,
if another thread (here setclientid) is quicker,
the reference to the confirmed record is tried to be reused
and that's where we end up in the endless loop.

The fix is to remove all references to the STALE clientid record.

Change-Id: I3dee5900bd732f11a4e2c66d35dfcfc282934b53
Signed-off-by: Swen Schillig <swen@vnet.ibm.com>
(cherry picked from commit 1f5844d66db24842e98ffd89f0bad019276290ff)