[fridgethr.c] Prevent infinite loop for timed out sync.
If a thread sync command was issued with a set timeout,
the while loop will be executed infinitely if the time-out pops.
In this case the callback is not triggered and therefore the
loop must be exited manually.

Change-Id: I422742eba9abba1c3278c0af5838e18d8541be80
Signed-off-by: Swen Schillig <swen@vnet.ibm.com>
(cherry picked from commit fd01c8c9fb1abfe96268e5652ee936b4e7c658c3)