MDCACHE - UP-call cleanups
- Set op_ctx

MDCACHE's up-calls need to call into MDCACHE internals, which assume
op_ctx is set correctly.  Rather than forcing the FSALs to set this, set
it at the MDCACHE up entry points.

- Only invalidate if cached

Don't create a new cache entry just to invalidate; instead, use the
internal APIs and return success if the entry is not in the cache.

Change-Id: Ic42c079b79c208ef4bb559428dec8dd2e6bfd6f8
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>