Fix state_lock refcount bug
Merging lock state removes merges from the existing entry into the new
entry.  If there is only one lock on a file and it is merged, then the
last lock is removed from the list, causing an extra refcount on the
object that is never freed.

Check for the list being empty before the merge to avoid this issue.

Change-Id: I86be1310025ebff7d93aadb50d872550eec2781f
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>