NFSv3 SETATTR should NOT set ATTR_SPACEUSED when ATTR_SIZE is set
ATTR_SPACEUSED is not a setable attribute and while it is related to
ATTR_SIZE, it is not the same (it's generally the number of blocks the
file occupies multiplied by the block size - i.e. an indication of the
actual space used on disk, and of course with sparse filesystems could
actually be LESS than ATTR_SIZE).

Change-Id: I5e62b46612cd0ec5c6ca51cd3165e38788c48d45
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
(cherry picked from commit 5a378c5dc52c93c1ddc6bc9b1c0595fa5cb13e06)