FSAL_CEPH: Use passed target obj handle type to decide unlink or rmdir
Since the API changed awhile ago to pass the target object of an unlink
(even if it wasn't cached before, it's cached now...), go ahead and look at
it's type to call unlink or rmdir.

NOTE: There is a very minor race situation where we could lookup and find
one object type, but between lookup and unlink/rmdir, it's been removed by
someone else and been replaced by an inode type incompatible with the type
of remove call we will make (non-directory replaced with directory or visa
versa). But that kind of race can happen if you do ls -l to determine
whether to rm or rmdir and then by the time you issue the command, someone
has done the same thing to you...

We COULD always still try an rmdir if we get EISDIR and unlink if we get
ENOTDIR... Or we could just leave them alone... On most of our clients,
the NFS REMOVE op originated from an unlink or an rmdir syscall, so they
already may have gotten it wrong and are subject to the same race.

Change-Id: Iea28a7bb4608c84564fb3ddeac86c4f11af4c3ae
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>