Release filesystem tree depth-first
Valgrind was complaining of accessing freed memory.  This was caused by
traversing the filesystem list in list order, sometimes freeing a parent
before it's children were freed.  When the child was freed, it tried to
unlink itself from the parent, accessing freed memory.

Instead, free the list using a depth-first tree walk, so that all
children are freed before their parents.

Change-Id: Ia6cc0a4ec1f0b814bbe9bd3cf441ebd1c20a8a55
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>