Chomp tailing slash from pseudopath
We already do this for fullpath, but we need to do this for pseudopath
as well to let the rest of the code not deal with it. For example,
pseudo_mount_export() wasn't dealing with trailing slash in pseudopath
causing missed junction detection.

Change-Id: Id136dd1e0cf0ea417c561ec441f8ab86c415c889
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
(cherry picked from commit 93631a9fb3612d4ddb848bbc44a2484980512da2)