FSAL_GLUSTER : for nfs-ganesha 2.4 gluster version should be atleast 3.8
The stackable mdcache was introduced in nfs-ganesha-2.4 . So there is a
possiblity for collision between functions in that layer with mdcache layer
in gluster client stack(their name convention is same). Kaleb send a patch
which avoids symbol collisions[1] in gluster, but it is available only from
glusterfs 3.8.

Also removing dependency over symlink mount, since only version of
glfs_h_lookupat() is available from gluster3.8 onwards.

[1] http://review.gluster.org/#/c/11814/

Change-Id: Ia4c86323d911593b24e1712f39f41c682602a3d8
Signed-off-by: Jiffin Tony Thottan <jthottan@redhat.com>