FSAL_CEPH: Set mask to GID when changing group ownership
Since we are changing the GID of a file we need to set the mask to
CEPH_SETATTR_GID so that we modify the right attribute of a file.

In the current situation the UID is set back to 0 without changing the GID.

Seems this was a small copy/paste bug which sneaked in.

Signed-off-by: Wido den Hollander <wido@widodh.nl>