Fix file handle bytes switching in xdr_nfs_fh4 XDR_ENCODE
FSAL_PROXY tests reveals a "Stale file handle" error at
first ls on root directory. This was due to call to htons
in the encode way that was switching bytes and corrupting
the file handle in FSAL_PROXY. To fix this issue, we simply
choose to let export id in network byte order in the handle
and only revert it when we extract it. This commit fix NFSv3
and NFSv4 protocols.

Change-Id: Ia682d0ce83d19ee873c1c7a993e6c86499c48bd7
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>