FSAL_VFS: Assure correct permission checking in open2
If open2 is by object, not by name, open2 does no permission checking.

If open2 is by name and was a create, the openat system call was made
with caller credentials and thus was permission checked. A NO_CREATE
call to just open the file was not permission checked, so assure that
the caller (open2_by_name) does a permission check.

Note that if FSAL_VFS is supporting ACLs, especially via the in memory
only version, the permission check done by openat may not be correct.

Change-Id: Ib20bfdc30a212f2755ac9efcb4bb4256dc8bc497
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>