FSAL_CEPH: Assure correct permission checking in open2
If open2 is by object, not by name, open2 does no permission checking.

If open2 is by name and was a create, the ceph_ll_create call was made
with caller credentials and thus was permission checked. A NO_CREATE
call to just open the file was not permission checked, so assure that
the caller (open2_by_name) does a permission check.

The way open2 is called, there is actually no functional difference
with this patch, however it should clarify things.

Also, a note is made that with the current ceph_ll_create interface
the alt groups list is not considered which may have an impact on
permission to create.

Change-Id: I829527cf9f3eeb23a5992dbded9feddf0f7863e9
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>