FSAL_GLUSTER: Assure correct permission checking in open2
If open2 is by object, not by name, open2 does no permission checking.

If open2 is by name and was a create, the gluster create call was made
with caller credentials and thus was permission checked. A NO_CREATE
call to just open the file was not permission checked, so assure that
the caller (open2_by_name) does a permission check.

A note is added to document that with a separate open call from the
create, there could be a permission failure in certain circumstances.

Change-Id: I7c962ed8e541b24d2071f53958f0a2c15b1f4fef
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>