FSAL_RGW: ssure correct permission checking in open2
If open2 is by object, not by name, open2 does no permission checking.

If open2 is by name and was a create, the rgw_create call was made
with caller credentials and thus was permission checked. A NO_CREATE
call to just open the file was not permission checked, so assure that
the caller (open2_by_name) does a permission check.

The way open2 is called, there is actually no functional difference
with this patch, however it should clarify things.

Change-Id: If904e44f89e2bca4bf3e82753d4ec72f8cf94347
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>