FSAL_GLUSTER: Use the new APIs to fetch & free upcall args
GlusterFS Upcall handling APIs have been remodelled to do
all the memory handling needed in gfapi and make the upcall
structure opaque so that in case if there are any changes
to it in future, it shall not break existing applications.

This change is to consume those APIs in FSAL_GLUSTER.

These changes are available in the libgfapi.so version >3.8.6
-https://bugzilla.redhat.com/show_bug.cgi?id=1347717.

Hence updated the library version needed as well.

Change-Id: Iad6c97a09aed801e7abd979cb452076bfef68b62
Signed-off-by: Soumya Koduri <skoduri@redhat.com>