Fix READDIR cookie traversal during deletes
If a READDIR is in process while deletes are happening on the same
directory, the return could be BAD_COOKIE when it should have been EOD.
This can happen in the case where all the files after the cookie were
deleted.

Another case is that if the file was deleted, but subsequent files
existed, we might have returned EOD instead of the next dirent.

Change-Id: I523f93f9d40b7c3a154548436ae3c8db19a77183
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>