MDCACHE - ensure that a handle has a parent pointer
If a NFSv3 mount of a sub-tree or a PUTFH with a handle not in the cache
occures, then the resulting handle won't have a parent pointer.  This
breaks lookup("..").

Ensure in lookup_path() and create_handle() that all handles have a
parent pointer (except the root of a filesystem, which is handled by
junction code)

Change-Id: I92c6cfd2ac3c14033837b138dfa3ac6527f101d7
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>