ceph: move readdirplus attribute handling into fsal_ceph_readdirplus
Add a return Inode pointer for passing back an inode reference and pass
that through to the new ceph_readdirplus_r.

Unfortunately, the legacy API wrapper relies on being able to do a
lookup to get attributes so we must also pass in a pointer to the
directory Inode, and a pointer to the creds.

Note that for now, we only pass in 0 for the flags. Eventually we may
want to change that to AT_ATTR_NO_SYNC in some cases, possibly via a
config file parameter? Passing in that flag may give better performance
at the expense of tight attribute cache coherency.

Change-Id: Ia9134f2cf60a2842996cb9552b8a6457d1d70ec5
Signed-off-by: Jeff Layton <jlayton@redhat.com>