ceph: convert FSAL_CEPH to new libcephfs abstraction layer
Now that we have a compatibility layer that can handle old and new
libcephfs APIs, wire the main FSAL_CEPH code to use it.

Note that in many cases, the Ceph FSAL was inappropriately passing in 0
as uid and gid. This changes the code to use the creds from op_ctx in
almost all cases.

Change-Id: I7fe3c842d80bf864bfce2401aa02fbed3e400e6e
Signed-off-by: Jeff Layton <jlayton@redhat.com>