cmake: shared lib FSALs should be MODULE libs, not SHARED libs
cmake documenation says that, for add_library(), a MODULE is a
flavor of shared library that you use dl_open()+dl_sym() on.

In general this is a no-op change, except that SuSE, on Leap42,
has added --no-undefined to CMAKE_SHARED_LINKER_FLAGS for its
rpm builds.

Unfortunately it has added --no-undefined to CMAKE_MODULE_LINKER_FLAGS
too; which I claim is a mistake. But one that can be worked around
without breaking their desired semantics for SHARED libs. But first
we need to make our modules into MODULES in cmake.

Change-Id: I907e7b6fb2076d2e5451f1959e609930047e85cd
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>