Fix nlm client refcount going up from zero
Once a refcount goes to zero, we should not be using it any more and let
the thread that made it zero free up the entry. If zero refcount entry
is found in the hash table, don't use it, remove it from the hash table
and insert a new nlm client.

Change-Id: Icda341d943cbd18bff4439e961748f6fda7f1092
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>