Fix NLM SHARE management for NFSv3 clients
When recording an NLM SHARE state for a file, do not set the
NFSv4 flag as this will erroneously cause denial of future NFSv3 writes
to that file. NLM SHARE is inherently an NFSv3 operation and as such
should not use an NFSv4 flag. This fixes the ability of Windows
clients to write to files when the “-o nolock” flag is not in use.

Change-Id: Id148431916cc381601a98634372e6c236553af41
Signed-off-by: Adam Kropelin <kropelin@amazon.com>