Fix multiple threads handling the same drc with zero ref count
nfs_dupreq_put_drc() drops the locks after decrementing the ref count to
zero. Imagine this thread not getting scheduled for a good amount of
time. The nfs_dupreq_get_drc() may bump up the ref count and recycle the
drc. The drc may eventually end up with zero ref count in another
thread. At this point we have two threads that are handling the same drc
with zero ref count. One thread may put it in the recycle list and the
drc could be freed before the other thread tries to access the drc! This
may lead to use after freed memory.

The fix here is to wait in the nfs_dupreq_get_drc until the first thread
puts the drc in the recycle queue.

Change-Id: I4d88bd401a5f266b4f7d75bfbeda9b91a329629b
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>