Restore attribute permission check in fsal_readdir
We check to see if the caller has execute permission in the directory
in order to allow fetching attributes, but we don't actually do
anything with that.

Move the fsal_readdir_cb_parms into the fsal_populate_cb_state and
set the attr_allowed based on the fsal_access check. Also just set
opaque once.

Change-Id: I7775db27155418ff5f4a0890a89f75f21567adf3
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>