Add config to allow NFS v3 and 9P mount by Pseudo instead of Path
A macro is introduced that selects fullpath or pseudopath based on how
the config param mount_path_pseudo is set. This is used to select the
correct path for 9P Attach, NFS V3 MOUNT, and NFS V3 EXPORT (which
displays the NFS V3 exports). It is also used in one NFS V3
specific log message.

Change-Id: Ia5cbb400a30455ce83dd2be7c1c4b72d2d13162c
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>