Info level logs to print compound op when there is an error
When there is an error in any othe ops in the compound op, no error is
printed in the nfs ganesha logs. The errors if any are printed only when
debug logging is enabled. But debug logging is very slow and we should get
the context of the failed op at inof level logging as well.

Change-Id: Ie54b78219ac87fe97c35201b51374100e81aa9df
Signed-off-by: Sriram Patil <spsrirampatil@gmail.com>