Logging too slow when full debug is enabled
Introduced a pool of files for logging. Every
thread will pick a file and log to it (sticks
to this file throughout). Threads log to a
buffer and leave. The buffer is periodically
flushed by the flusher thread or by the current
thread if the log is full. Significant response time
improvement seen in case of mount, ls -l, touch
of many files, etc.

Change-Id: Iefa8b8d48d2b29c68e9df6347a81eef932efe39f
Signed-off-by: Satya G S <g.satyaprakash@gmail.com>