Closing the log file preiodically so that logrotate works well if used
If logrotate is used with nfs ganesha logs, the log statements are appended
to the rotated file if it is not closed, because the inode does not change.
So making sure that the log files are closed periodically in order to
create new ones in case log rotation is used.

Change-Id: I79bad8ae1c870c69780d306c462cd6604c741592
Signed-off-by: Sriram Patil <spsrirampatil@gmail.com>