AVL: Make inlining easier and allow external to break up lookup/insert
Everyone wants AVL access to be inlined, make it possible to have
avltree_insert entirely inlined.

There is at least one use case for a function to do an avltree_lookup,
and if the key is not found, subsequently add it and since avltree_insert
does an avltree_lookup, we just separate out the rest of avltree_insert as
avltree_do_insert.

Change-Id: I32f7b6f3a884958a7231ed03695d6c840c1879d6
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>