Fix nsm client refcount going up from zero
Once a refcount goes to zero, we should not be using it any more
and let the thread that made it zero free up the entry. If zero
refcount entry is found in the hash table, don't use it, remove
it from the hash table and insert a new nsm client.

Change-Id: I68a4e848a13981932e1d34dc654de5985f3fb491
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>