Fix nlm state refcount going up from zero
Once a refcount goes to zero, we should not be using it any more
and let the thread that made it zero free up the entry. If zero
refcount entry is found in the hash table, don't use it, remove
it from the hash table and insert a new nlm state.

Also, under some cases, we do delete non zero refcnt nlm state objects
from the hash table as well. This change should avoid logging spurious
messages from dec_nlm_state_ref() for not finding the nlm state in the
hash table and accidentally deleting someone else's nlm state.

Change-Id: I3917327f2fdfe8df0e5d9d2a64786ba4c1eb4cd4
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>