FSAL_GLUSTER: Set caller credentials for backend operations
When nfs-ganesha service is run as non-root user, unlike root
user that particular user may not have permissions to do certain
operations. Hence pass actual caller-credentials to the backend
for permission check.

getattrs is the only operation where we are not switching to
caller credentials, as every user should have permission to
read stat and POSIX_ACLs. This needs to be revisited when we
support RICH_ACLs as they may restrict READ_ACL to particular
users. Frank suggested that we may then should not do permission check
in ganesha layer but offload it to backend.

Change-Id: I318973b206e83835d55454a785836853139fdd83
Signed-off-by: Soumya Koduri <skoduri@redhat.com>