.gitignore: allow debian .patch files
The Ubuntu packaging layout with git-buildpackage assumes a
"debian/patches/" directory with several .patch files in it.

When upstream's .gitignore tells Git to ignore .patch files, we have to
edit that line out downstream. When we forget to do that downstream, it
can lead to missing patches and broken downstream builds.

Allow patches in the /debian/patches directory so it's easier to
maintain an Ubuntu package based on upstream's Git repo.

Change-Id: I3d7ed0af45ba3a18a7da9363b8ed339d4ce5460d
Signed-off-by: Ken Dreyer <kdreyer@redhat.com>