Determine size of ACLs to be encoded when getting ACLs
If nfs4_getfacl is run by the client when
there are large number of ACLs and if the server is not
able to fit in all ACLs in the buffer of size
NFS4_ATTRVALS_BUFFLEN then the client gets Input/Output
error.
To avoid this we now calculate at run time the size
required by ACLs, to determine the size of buffer where
the ACLs would be encoded.

Change-Id: I4ace8223abe2c2957f6e40cb619a1b838ca78677
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>