GPFS: Use "obj_lock" when closing fd.
Currently gpfs_close() does not use lock when closing a
file descriptor. Thus when the same file is removed by two
clients at the same time, there is a chance to see a
CRIT message with error ERR_FSAL_NOT_OPENED in the log for
one of the clients.
To fix this we now close the file descriptor under the lock
"obj_lock".

Change-Id: I6e5f74e2d6985b8c5952c18021cf296ae291bc7b
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>