[FSAL_GPFS] remove fsync() from gpfs_write2()
We don't check for EUNATCH error after fsync() system call, so we end up
propagating a failure to NFS client applications.  GPFSFSAL_write
or gpfs_write_plus_fd take the given fsal_stable and sync the data if
needed. So there is no need to make a separate fsync() here.

Change-Id: I4ff7e14f3323bcc4f0dc3fc3a4e21435992ea308
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>