MDCACHE: mdc_lookup_uncached needs content_lock held for write
Fix one call in mdcache_readdir_chunked.

Note that one call in mdc_lookup is for when we aren't caching dirents
in the directory, so mdc_lookup_uncached will never do anything that
needs the write lock.

Note that in legacy dirent cache, mdcache_readdir would need to drop
the read lock and acquire the write lock, and this would require
significant coding. Since we are deprecating the legacy cache, we
hope this actually doesn't come into play (but we've taken note
such that if an issue actuallt arises, we can try and fix it...).

Signed-off-by: Dominique Martinet <dominique.martinet@cea.fr>

There is complexity here because dropping the content_lock means the
state of the chunks and dirents may have been changed.

Change-Id: If2e656809d94bd5628eef26ed5529b212ec50a41
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>