Fix a race in using hashtables leading to crashes
get_state_owner, get_nsm_client, get_nlm_client and get_nlm_state
functions use atomic_fetch_int32_t() on the refcount to see if their
hashtable entries are about to be deleted. If not, they later place a
hold and use the entry! It is possible that an entry might have a
positive refcount when atomic_fetch_int32_t() was used, but it could go
to zero by the time these functions place a hold on the entry. In other
words, these functions end up using entries that are about to be deleted
and freed by some other threads leading to ganesha crashes.

Change-Id: Ife9c8b6c5247ae9c902c4259fb8a8c11dc512477
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>