Set op_ctx when calling mdcache_lru_clean() in the context of background thread.
The background thread goes through each lane to find entries to be cleaned.
In one scenario, it resets the op_ctx to NULL which results in the assert()
to be hit in mdcache_lru_clean(). This fix moves the reset of op_ctx to NULL
after the call to mdcache_lru_unref().

Change-Id: I0896e9711c3c09449df62a58698d622739ca18da
Signed-off-by: Pradeep <pradeepthomas@gmail.com>