Create v4.1+ openowner with confirmed flag already set
With the introduction of sessions, OPEN_CONFIRM is now a no-op in
v4.1+. Go ahead and set the flag to mark an openowner confirmed when
it's associated with a v4.1+ client. This allows us to eliminate some
minorversion special casing as well.

Change-Id: I382082dda3a90ceebae50b30f13500b0a93717dc
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 7934673a549e4355f0a008c8db544f5825e73629)