GPFS: disable delegations temporarily
The core delegation code has been pretty broken for a long time now, so
we're reworking it into something saner. Once we do that, GPFS will need
to be converted to use the new interfaces. In the meantime, ensure that
GPFS won't try to get a delegation by disabling it in the FSAL's
staticinfo.

Change-Id: I82b78b4dfc36601e8fdd74cb91e47b8894c20635
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 1e88dbef583e9eabec4f79659badcfc461071452)