NFSv4: drop DO_DELEGATION define
I don't think we need the delegation code commented out. There are
enough configuration settings that nothing should really have this on
now anyway, and anything that does (GPFS?) should basically work with
the other recent changes that have gone in.

Also, just remove state_open_deleg_conflict. The counters for this live
in the FSAL now, so it can't work anyway. The FSAL must enforce this if
necessary.

Change-Id: I92c32f142b211ea9f1f9fb0b27c39c2de57fc257
Signed-off-by: Jeff Layton <jlayton@redhat.com>