Fix dec_client_record_ref accessing freed memory
Let us say thread1 in dec_client_record_ref() decremented the refcount
and found it to be 0. It is ready to delete from the hashtable but got
scheduled out before acquiring the hash latch!

Another thread, thread2, that needs this record, found the record in the
hash table and incremented the refcount from 0 to 1.

Also Assume that thread2 is done with the record and eventually called
dec_client_record_ref() to release it. It decremented the refcount to
zero. Now we have 2 threads, both finding the refcount as 0!

One of the threads will win to acquire the hash latch and actually find
the entry in the hash and free it. The other will use the freed memory
to find the entry in the hash table! Worse yet, it may find a record
inserted by some other thread (not its own) and free it!

Fixed so that the refcount never goes up from 0 like other nlm/nsm hash
tables.

Change-Id: Iafcd46efa70c86c4ae0bc4d289e03b11c6b6f95d
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
(cherry picked from commit 7d0629b3bc647aa897eb70705858e5759b9911a6)