NFS4 - Don't leak ref in error cases of LOOKUPP
If we try to look up the parent of the root of an export, any of the
error cases (including that root being the top of the tree) were leaking
the reference on that root object.

Release this reference in all error cases.

Change-Id: I2506d830eaa7b21aa57e9026e0dea78d73fde07b
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit d7ce4c4c101291c1a325ae97116a45f54d930c37)