mdcache: don't hold the attr_lock for write over the setattr2 call
This is problematic, as some FSALs (e.g. Ceph) can take quite some time
to do their work. During this, getattr activity is stuck, and that can
lead to some deadlock situations with delegation returns.

Do the setattr2 call without holding the mutex, and only take it to
update the cache.

Change-Id: I08e2e752357e76d77583bd22f15f8bd72d71f39e
Signed-off-by: Jeff Layton <jlayton@redhat.com>