mdcache: don't fail setattr if it succeeded but getattr failed
We call down into the fsal to do the setattr, and after that succeeds
we do a getattr to update the cache. If that getattr fails, we fail the
setattr, even though the fsal said it succeeded.

Change it to return success in that case, as the setattr supposedly
did succeed, even though the getattr failed. In the event that the
getattr fails, invalidate the cache entry appropriately.

Change-Id: I52330c2fcb16ca469068f5a66258e21232db1baa
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 6fa5412cda62dad5d1c649602bcde7c9245141a0)