SAL: allow grace period to be lifted early if all clients have sent RECLAIM_COMPLETE
Keep track of the number of clients on the recovery list, and the number
on the list that have sent a RECLAIM_COMPLETE.

v4.0 clients will never send a RECLAIM_COMPLETE, so when those values are
equal we know we have all v4.1 clients that have finished reclaiming and
so can lift the grace period immediately.

Since we don't have a great way to detect whether any SM_NOTIFY requests
were sent in conjunction with this restart, we'll only lift the grace
period early if NLM is completely disabled (a'la Enable_NLM setting).

Change-Id: Ie720da438220f7115d4458b4b995753d480d6491
Signed-off-by: Jeff Layton <jlayton@redhat.com>