fsal_create() - Add case for normal files
Traditionally, fsal_create() only created non-normal-files, and asserted
if an attempt was made to create a normal file.  The only way to create
a normal file was with open2().

The ability to create a file without tracking it's open state is very
useful to the unit test framework, so add this ability to fsal_create(),
by opening and then closing the file.

Change-Id: I31badf93e2505f07adf5dfbe8605645081f31f95
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>