pNFS - Fix memory leaks for fsal_pnfs_ds
Instances of fsal_pnfs_ds can be of 3 varieties: 1) Associated with an
export (MDS/DS co-located).  This will be cleaned up when the export is
cleaned up.  2) Stand-alone DS (No MDS).  This was leaking.  3)
Placeholder when no DS is configured.  This was leaking.

Fix case 2) by tracking all DSs and cleaning them up on shutdown, like
is done for exports.

Fix case 3) by detecting it and using a static instance, rather than a
dynamically allocated instance.  This needs to be done because the
placeholder instance is never committed, so it's otherwise completely
untracked.

Change-Id: Id66c8c3c9056a69c5306fca6895c66782985b1a5
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>