Fixing NFSv4 referrals in FSAL_VFS
FSAL_VFS was returning incorrect fs_root as part of FS_LOCATIONS reply.
FSAL_VFS should return the referral directory path as fs_root, but it must
be a pseudo path. FSAL_VFS was returning the actual path. When the Path and
Pseudo are different for an export, referrals fails with ENOENT.

This fix makes sure that we replace Path woth Pseudo path in the fs_root
path.

Change-Id: I7e570470226bb1ebff8ceca2ab4062f52fc4fe79
Signed-off-by: Sriram Patil <sriramp@vmware.com>