Let idle decoder threads exit!
We create a decoder thread per each xprt (connection). Due to a bug in
fridgethr_freeze() for type fridgethr_flavor_worker threads, they never
exit even after closing the xprt!

The fix is to let the code go to restart label and return when there is
nothing left to do.

Change-Id: Ic1bf2808316ea17e398b7053c121ae36fadc2f1c
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>