move read EOF handling to fsal_read2
ESXi requires RFC-compliant use of the EOF flag in read3 calls:
The EOF flag must be set if the EOF was reached by the read.
Most FSALs only set EOF if a read at the EOF was attempted,
i.e., a read returning 0 bytes because the offset was >= file size.

This was previously only handled by the NFS4 implementation,
breaking compatibility with ESXi NFS3 clients. This patch
moves the logic up one layer to handle it for all protocols.

fixes #177

Change-Id: I83c7f1fae91054b32af7ca90f56a0ff064a05cfe
Signed-off-by: Paul Emmerich <paul.emmerich@croit.io>