build: compile conf_lex.c with _GNU_SOURCE to get caddr_t definition
Build fails on prerelease of Ubuntu 18.04 (bionic) because caddr_t is
no longer implicitly defined when <sys/types.h> is #included.

Trying to untangle the #ifdefs and #ifndefs in <features.h> and
<sys/types.h> (which is /usr/include/sys/types.h on redhatish linux
and /usr/include/{x86_64-linux-gnu,...}/sys/types on debianish linux)
it is revealed that caddr_t is defined when __USE_MISC #defined.

And __USE_MISC is #defined when either _DEFAULT_SOURCE is defined or
_GNU_SOURCE is defined. (_DEFAULT_SOURCE is/was a replacement for
_BSD_SOURCE and/or SVID_SOURCE which were deprecated in g(nu)libc some
time ago.)

Some combination of distributions implicitly or explicitly enabling
newer and stricter ANSI C compilation has resulted, in the case of
Ubuntu 18.04 anyway, caddr_t not being defined.

Change-Id: Ie8eb79656fc5c3c48073b42f73bdcb92614b7c7c
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>