nfsd: handle rpcbind registration errors more gracefully
Registering with portmapper/rpcbind is optional for v4 servers. Failure
to do so should not be a fatal error.

Change the code to return a boolean (false when registration fails), and
fix up the caller to ignore v4 registration errors.

Note that there is also another behavior change with this patch: do not
register mountd ports unless v3 is enabled.

Change-Id: I2002660b2cfab3659d6d8daa2fe2a13135510706
Signed-off-by: Jeff Layton <jlayton@redhat.com>