python: prepare for python2->python3, explicit python2 in shebang
Python2 will reach EOL in 2020. Red Hat has announced that the next
major version of RHEL (8?) will use python3 by default. We can only
guess what python2's status will be. Possibly deprecated, perhaps
available but strongly discouraged, maybe even completely unavailable.

For Fedora (and RHEL, CentOS, Sci Linux) at least, we're not supposed
to use '#!/usr/bin/env python', see
 https://fedoraproject.org/wiki/Packaging:Guidelines?rd=Packaging/Guidelines#Shebang_lines

2) we're strongly suggested to not use '!/usr/bin/python', see
 https://fedoraproject.org/wiki/Changes/Avoid_usr_bin_python_in_RPM_Build#Quick_Opt-Out
We can use a magic environment variable when building RPMs to delay the
inevitable for this. (But I believe many would prefer not to to put it off.)

The eventual goal is that eventually everything is made explicitly python3
in the shebangs and in any build and run-time invocations of python.

Change-Id: I4f6c529a1dcbb67f99dc57cde199fd13a58a4ea4
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>