rados_grace: don't create grace db on startup
The current rados_grace design requires that the nodeid be present in
the grace db before it will allow it to join the cluster. That means
that the DB must be present before we start the server, so there is
no need to have the server create the database here.

Change-Id: I595714f50dffb4ab00e0f1d7398d7935f96d5d65
Signed-off-by: Jeff Layton <jlayton@redhat.com>