rados_cluster: start grace period after creating new recovery DB
We don't want to mark ourselves as enforcing until we have the new
recovery DB in place. Another node could end up lifting the grace period
before we've generated the new database. An inopportune crash at that
point could lead to client records being lost in the transition.

Start the new grace period after changing the OID strings and creating
the database for the new epoch.

Change-Id: I628537d75856b1549bc578de8428bec484bd2cbd
Reported-by: J. Bruce Fields <bfields@fieldses.org>
Signed-off-by: Jeff Layton <jlayton@redhat.com>