Revert dirent accounting
c7c336e1cd2495ec5ffbb0dac7de31ac5711c077 changed dirent accounting to
track the number of dirents.  However, with 128 dirents / chunk, and
with chunks in the LRU, it could happen that a readdir reaped chunks
from that readdir.  This change is still desirable, but needs to be
reworked with protection for in-use chunks.

Change-Id: I6b3110767e566bb13ebef39896857d71a9583e17
Signed-off-by: Daniel Gryniewicz <dang@fprintf.net>