FSAL_VFS: fix dummy handles on vanilla FreeBSD
We cannot make use of fid_reserved like on PanFS since it may be used
internally by other filesystems. Introduce a new fh_flags field to
v_fhandle and make use of it to flag dummy handles.
We now need to use the v_to_fhandle macro to pass file handles to kernel.

Change-Id: I0e64f29f879ca56093a7631663384472a50cd3b8
Signed-off-by: Fatih Acar <fatih.acar@gandi.net>