build: cleanup unused defines in config.h
Many defines in config.h were left-over from the old configure script,
where the file was autogenerated, but not actually used.

Having less defines reduce the maintenance burden (e.g. not having to
rename it!) so unused macros should be removed.

These were found and manually checked with this command:
 awk '/cmakedefine/ {print $2}' < src/include/config-h.in.cmake | \
   while read -r v; do
     echo "--------------------------"
     echo "$v"
     git grep -w "$v"
   done | less

Change-Id: If96ff7bd41f395d90bd647194f4de0f7bb5e2d74
Signed-off-by: Dominique Martinet <dominique.martinet@cea.fr>