FSAL_CEPH: kill off old session before the mount
One problem with running ganesha over cephfs is that the MDS will
preserve old sessions and their state until the old client times out.
This means long stalls for the new server taking over after failover.

To help address this ceph has recently grown some new interfaces to
allow its clients to take over state held by previous incarnations of
themselves. For now, we can only label a session with an opaque string,
and request that older sessions with the same label be kicked out.

Add a cmake test for the new cephfs reclaim infrastructure. If it's
present, set the uuid for the session to the cluster nodeid, and request
that any previous sessions be reset when establishing a new session.
This allows for faster takeover from old server.

Eventually the new interface may grow the ability to seamlessly take
over state held by the previous client, which would obviate the need for
a grace period on failover.

Change-Id: I70e93cbff2e2ae80989aaa9ee9c96d3031f5e919
Signed-off-by: Jeff Layton <jlayton@redhat.com>