fs_locations: split location field into server and rootpath strings
Sriram did some work a few months ago to plumb in a new vfs-layer
mechanism for fetching fs_locations info. Unfortunately, it ended up
squashing the "server" and "rootpath" fields into a single string,
delimited by a ':'.

Presumably this was done because the xattr presents this info in that
format, as later in the generic code, that string is split into two for
encoding the fs_locations attr on the wire.

In later patches we're going to want to make this more suitable for
generating fs_locations_info, and keeping them conjoined like this will
be more problematic for that use later.

Split the location string back into two in fsal_fs_locations_t, and
just have FSAL_VFS concatenate a new ':' delimited string when it
needs to generate one for the fsid hash.

Change-Id: If3ea29d6c0053b8aeb4644c26dc46c2c12b3e88d
Signed-off-by: Jeff Layton <jlayton@redhat.com>