NLM: don't check for grace period in nlm4_Unlock
There is no need to check for the grace period when processing an
unlock. The grace period exists to prevent clients from acquiring
conflicting state that other clients have yet to reclaim.

An unlock, however, can't cause a conflict since we're releasing state
that is genuinely held by the client.

Change-Id: Iad2f36b9579f868d7715636f43d1c2ce2479a12b
Signed-off-by: Jeff Layton <jlayton@redhat.com>