NFS: don't check for grace period in nfs4_chk_clid
We will validate cid_allow_reclaim flag at the point where we are
actually reclaiming. There are also some ways that we can go back into
grace and load a new recovery db after coming out of it with certain
dbus commands.

In that situation we could conceivably have a client establish a session
outside of the grace period and then issue reclaim requests against it
after the grace period has restarted (maybe a merging of state held by
two servers in the cluster?).

Drop this check from the code since we always check cid_allow_reclaim
in conjunction with grace period checks at the time of the reclaim.

Also, eliminate a comment in nfs4_chk_clid_impl. We definitely do not
want to remove the client, as we could end up having to reestablish the
clientid for some reason, and removing it from the list would render it
unable to reclaim anything.

Change-Id: I7cfd5ad6cd74c9358e9ec86710d0a31ad1ffa883
Signed-off-by: Jeff Layton <jlayton@redhat.com>