mdcache: do readdir cb supercall using entry->attrs
mdc_readdir_uncached_cb will update the mdcache first, and then do the
supercall with the old attributes. When updating the cache though, we
might "steal" some of the pointers in "attrs", which could make us
populate the reply from an incomplete set of attrs. Use the attrs in
the mdcache instead.

Change-Id: If2e3b7e9de1f496c96ad2cb29552318e33626c61
Signed-off-by: Jeff Layton <jlayton@redhat.com>