ganesha-rados-grace: only create pool and db on "add"
The first thing we have to do with a new cluster is add members, so it
doesn't make sense to create the pool or db unless we're doing that.
Otherwise, the tool will create the pool and db, even if we're just
dumping the current status.

Also, clean up some unnecessary code duplication. We need at least one
nodeid for all commands except for "dump" so just check that after we
see if we're handling a dump command.

Change-Id: Id6f72384a0fb4ae3e7286cad71fa8b8fda90607f
Signed-off-by: Jeff Layton <jlayton@redhat.com>