readdir_correctness - Empty cache during run
Even though we drop our refs, the way that the entries are created
causes them all to stay in the cache, even though reachable.

Add a debug function to MDCACHE to allow cleaning up all reapable
entries in the cache, so that readdir will trigger all the "missing
entry" cases.

Change-Id: I99f6bc9d5ae2bcf354d4eb2f6c510f9926d85f97
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
Signed-off-by: Daniel Gryniewicz <dang@fprintf.net>