SAL: decrement reclaim_completes counter when expiring client
A client could send a RECLAIM_COMPLETE and then end up reestablishing
the clientid during the grace period, and then send another one. If that
happens we'll expire the old client, but the reclaim_completes counter
will be artificially high.

Ensure we decrement reclaim_completes when expiring a client that has
sent one.

Change-Id: I046a95e320c9fed2ae887eab3056f95ff4ab7951
Signed-off-by: Jeff Layton <jlayton@redhat.com>