FSAL_CEPH: add FS cluster ID into the FSAL_CEPH filehandle
Currently, the FSAL-specific portion of the filehandle is composed of a
vinodeno_t, which is just an inode number and snapid tuple.

In a follow-on patch we're going to allow exporting alternate, named
filesystems within the same cluster. At that point, the vinodeno_t is no
longer unique enough for this purpose.

This patch expands the ceph_handle with a new 64-bit fscid field, that
denotes the cluster ID within the filesystem. From there, teach the
cluster to encode the fscid in the handle when it's available.

Ceph reserves fscid 0 to denote "legacy" filesystems. If no filesystem
name string is specified, or if it is a legacy filesystem, then we don't
append the fscid to the handle, in order to maintain backward
compatibility.

Change-Id: Id8818f181f9abeb49b2845c551e6e2b83e80db9b
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit b3d97f8157a131f55d848ff57e46af91b746b944)