Add nfs_req_result type and convert nfs4 compound ops to return it
Instead of returning the redundant nfsstat4, return nfs_req_result.

We add a new value also NFS_REQ_ERROR (which is used by nfs4_Compound
to determine if it should stop the loop).

This will be more meaninfull once we add a nfs_req_result to indicate
a need to suspend an async request.

Change-Id: I05029037bf09e32a678cfaf56b74e47d60cb62f9
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>