exports: prune off old exports after reloading config
After rereading the exports in a new config, walk the list of them and
prune off any that have a config generation older than the one
representing the current config.

If pruning an export would leave another disconnected however, leave it
in place and LogCrit a message indicating that it couldn't be unmounted.

Change-Id: If5b5a11c7b45150ec34d7bc7b500c341f9126f79
Signed-off-by: Jeff Layton <jlayton@redhat.com>