MDCACHE - Don't return dead entries from hashtable
When the refcount of an entry goes to zero, it's removed from the LRU
under the lane lock, and then the lane lock is dropped, and the entry is
cleaned up.  This includes removing it from the hash table.

However, there is a window during which mdcache_find_keyed_reason() or
mdcache_new_entry() can find it from the hash table, and increment it's
refcount to 1.  This causes it's refcount to go to zero again, recursing
into cih_remove_checked(), causing an assert.

Fix the hash table to not return an entry if it's refcount is zero.

Change-Id: I9a922a1fb54561f04d71cc83c0bcd3289ec854ae
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit 25320e6544f6c5a045f20c51446f57c9dc036412)