rados_urls: set up watch if RADOS_URLS block configures it
We can send ganesha a SIGHUP to tell it to reread its configuration. In
a containerized environment however, sending such a signal represents a
potential security issue as we need to open up permissions to allow it.

This patch adds a new watch_url option to the RADOS_URLS config block.
When that points to a valid RADOS object, set a watch on that object.
When we get a notification on that object, send a SIGHUP to ourselves
to force a configuration reload.

Change-Id: Ia83329eb0774f8b0bff428b60fc72df1a70b47c8
Signed-off-by: Jeff Layton <jlayton@redhat.com>