Do not refresh attrs if the file is delegated
If a file is read delegated, then no other client can
modify the file data and hence we can trust the cached
attrs. And incase any client holds write delegation, we either
update cached attributes via CB_GETATTR or recall
the delegation before responding to the other clients.

Hence it seems safe to avoid refreshing regular attributes
from the backend FSAL if there is delegation granted to
a file. However this should be the case only if the md-cache
is enabled (i.e, expire_time_attr is not set to '0').

Change-Id: Iffbe76b6c089b6435826a7f72240f9a9804e5ddd
Signed-off-by: Soumya Koduri <skoduri@redhat.com>