MDCACHE - Fix dirent chunk linking
We can't depend on the order of the chunk list for finding a previous
chunk.  This is for several reasons.  1. chunks are reaped out of the
middle of the list. 2. A readdir can start on a missing chunk, meaning
there's no way to find the previous chunk, even if it exists.

Remove the ordered aspect of the list, since we can't depend on it, and
keep a prev_chunk pointer in the state.  This allows us to set next_ck
going forward from whatever chunk we start with.  There will be fewer
cases when next_ck is set correctly, but we won't end up with cycles
where next_ck points further back into the directory.

Change-Id: I9d48c2937f2deaf3dd3614da6b9e1df1fd5e67e3
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit b8fe6364c61c4ac0c086c67b4d06685beb55bc5f)