rpm: fix selinux policy compression paths
When we compress the ganesha policy file, we are already in the top
source directory, so the "selinux" directory is directly in the current
working directory. Update the pushd command to reflect this.

The policy file we create is "ganesha.pp", not "nfs-ganesha.pp". Update
the bzip2 command to reflect this.

Change-Id: I2184fcfbcc8c58a0a65123a57b0586b17a0ef5cc
Signed-off-by: Ken Dreyer <kdreyer@redhat.com>