MDCACHE - Don't drop the sentinal ref with the lock held
Dropping the sentinal ref in cih_remove_checked() may cause a cleanup
of the entry, which will recurse into cih_remove_checked() again.  This
will abort, because the lock is already held by this thread.

Instead, drop the lock before we actually drop the ref.  That way, if
it's the last ref (which is common in error cases), the cleanup will
recurse, and not find it in the hash table.

Change-Id: Iccf982f2b2d85a3617f1604e0636afce9501fdfc
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>