FSAL_GLUSTER: Fix a fd ref leak in lock_op2
In glusterfs_lock_op2(), if the associated open_state has a
open fd, we dup the fd using glfs_dup and re-use it.

glfs_dup() allocates new glfd object which takes ref on
the underlying fd maintained by gfapi stack. This needs to be
released post processing the fop. Otherwise it leads to
fd leak. This patch addresses the same.

Change-Id: I264e4e957791ea88caaba9e7f3ae85157ce2dcbf
Signed-off-by: Soumya Koduri <skoduri@redhat.com>