MDCACHE - Ensure ATTR_ACL bit matches attrs.acl ptr
The refresh/copy/ref-handoff code depends on the ATTR_ACL bit in
request_mask always matching the attrs.acl pointer.  However,
mdcache_refresh_attrs() overwrites request_mask with the given one,
losing this relationship.  Fix it up, to make sure the bit is always set
when the pointer is non-NULL.

Change-Id: I15c556e806184ba42ed41519618f74a8aa66c6df
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>