MDCACHE - Fix mis-set of first ck in directory
After the unchaining, we set look_ck to 0 when we don't know the next
chunk.  This triggered the setting of first ck for the directory in the
middle, skipping whole chunks of the directory.  Make sure we skip this
when look_ck is zero.

Change-Id: Ie14ecac3d41f6fb6cf2ee95d86a2b41e369601fb
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit 7558806388adb9996fa02994b45dd152c5cd53d8)