Preserve old ACLS
"fsal_copy_attrs" function copies the ACLs based on the request mask.
In case of retaining old ACLs, we have to explicitely set the mask so
that old ACLs are preserved. And later we have to unset the mask to
avoid updating acl_time in case of preserving the old ones.

Signed-off-by: Sachin Punadikar <psachin@in.ibm.com>

Change-Id: Ib45bd2004d93b9bc945c1cb3724f8c574a5c37a0
Signed-off-by: Amit Sutar <Amit.sutar@ibm.com>