IBM-only: Use setfsgid/setfsuid instead of setresgid/setresuid
2.7 code uses setresgid()/setresuid() to change effective gid/uid.
But this is not working as of now as uid is not changing with
setresuid().
Temporary fix: Use setfsgid()/setfsuid() instead of setresgid/setresuid()
in the way it is used in 2.5. Will revert it when we have a fix to
use setresuid() to change uid.

Change-Id: Icff0a50e168a02487885273e2c59f303d4634ae3
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>
(cherry picked from commit 41749b2e5ad50d1e5f94366ae61809f180e849d6)