NFS4 - Allow locking past the FSAL's maxfilesize
rfc5661 only allows lock failers when the offset + length overflows a
64-bit integer.  However, many FSALs have a max file size of 63-bits.
We were returning BAD_RANGE in this case.

It seems better to allow this, so fix it up so that an attempt to lock
past maxfilesize will be transformed into "lock the rest of the file".

Change-Id: I634ea27cc2dfe22e91ace0ddb9027d2bc36906c6
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit 3d069bffc10bbf5538365a3f145236e0dae118a7)