FSAL_MEM - Allow whence_is_name readdir
Change the index tree to use a hash of the name, rather than a
sequential counter.  This makes the dir order more like a real FS, and
less likely to hide bugs.  In addition, it allows a traversal order tied
to the name.

Add a global that allows setting whence_is_name for MEM.  No other
changes are necessary.  This allows testing the MDCACHE whence_is_name
code.

Change-Id: Ic20fe9aa43316ecbf3eec4e1943354ac4efed21c
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>