MDCACHE - Restart readdir if directory is invalidated
A lookup of an entry can nuke a dirent if the keys change for the entry.
If the new dirent can't be placed, then the directory will be
invalidated.  In this case, we need to restart the readdir, picking up
from where we left off.

Change-Id: I73344e9cde2bdccfef6566d0e2b3a6675de00a9c
Signed-off-by: Daniel Gryniewicz <dang@fprintf.net>
(cherry picked from commit fdaebeec28365494bb571a53dd7d6a071b59736a)