nfs_ip_name cache: Larger hostname buffer and other fixes
getnameinfo specifies using NI_MAXHOST as the size that should be used
for the hostname buffer.

Just use the caller's buffer to getnameinfo into to save a copy.

Make the hostname dynamic sized in nfs_ip_name_t

Even if there is a hash error, we still resolved the ipaddress into
a hostname or didn't and are going to just use the ipaddress as the name,
so no reason to fail to return what would have been cached to the caller.

Make sure all the string copies are safe and report an error if the buffers
are too small.

Change-Id: I9f5daaf632026dce9882003260e43bd17ae5fc25
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
(cherry picked from commit dc2c0f1628aadcfd33e880f1abdad5d5bc84c833)