Clean up rpc_tools.c including fixingup sprint_sockip etc.
Eliminate sprint_sockaddr and have callers directly call
display_sockaddr.

sprint_sockip is ONLY called for AF_INET and AF_INET6 (and maybe VSOCK
in the future) so we can be a little simpler.

Make sure all callers of sprint_sockip are ok with possible failure.

Call display_sockaddr instead of sprint_sockip if the use really is
just for logging.

Change-Id: I08278ba0230f79936094c01cd1414db13290b94b
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>