MDCACHE - Hold lock while dropping ref on chunk
Dropping the ref on a chunk requires holding the content lock, and
requires holding it for write when the last ref is dropped.  This ref
won't be the last, but we still have to hold the lock for at least read
to ensure another thread doesn't drop a ref making it last.

Change-Id: I1a268d4ec8de23d16be01b317c2118f3098bbefa
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit e306c01685b74970ada2259f2bf393cdc1233af6)