mdcache_get_parent: Don't do lookup of parent for a root object
Currently mdcache_get_parent() may get called for an 'entry'
where 'entry->obj_handle' points to the root object. In that
case we would be doing lookup of parent for the root, which
we should not be doing.
To avoid this prior to doing lookup check if the
entry->obj_handle points to a root object, if yes then don't do
lookup of parent for the 'entry'.

Change-Id: Ibf25fd283573c9fa28ae0452d96f00f294f66828
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>