ganesha-rados-grace: add --cephconf and --userid command-line options
On a machine that may act as client to more than one cluster, we need a
way to tell the program what config file to use. Add a new --cephconf
command-line option.

Also, we may need to connect as a specific cephx user. Add a --userid
option to allow for that.

Change-Id: I149f48ee18546bcd2be3bb8536a91ca7451af1ff
Signed-off-by: Jeff Layton <jlayton@redhat.com>